new full constructor for dir. few warnings squelched. working dirMultiNaiveBayes with cat data

+++ include/dirBaseMeasure.hpp
@@ -47,7 +47,7 @@ public:
   void fromMerge(const DirSampled<Disc,T>& dirA, const DirSampled<Disc,T>& dirB);
 
   void print() const;
-  virtual uint32_t getDim() const {return(dir0_.alpha_.size());};
+  virtual uint32_t getDim() const {return(uint32_t(dir0_.alpha_.size()));};
 
 //  const Matrix<T,Dynamic,Dynamic>& scatter() const {return dir0_.scatter();};
 //  const Matrix<T,Dynamic,1>& mean() const {return dir0_.mean();};
